move private entrypoint helpers to singular module
